Only queue inspections for group unit IDs.
Use a method from MooUnit-1.0.5 to test that a unit from a
MooUnit_UnitChanged event is actually a group unit ID before
queueing an inspection for it.  This makes it so if a healer
targets a group member (partyNtarget), an inspection is not queued
for the group member.  The change greatly reduces the number of
unnecessary inspections.